perf(cli): optimize status command startup by breaking heavy import chains
Break 3 unnecessary transitive import chains that caused the status
command to load AJV, AWS SDK, baileys, jiti, and babel at startup:

- Extract PROTOCOL_VERSION to gateway/protocol/version.ts (breaks AJV
  100-validator compilation chain from gateway/call.ts)
- Dynamic import of @aws-sdk/client-bedrock in bedrock-discovery.ts
  (only needed when discovering models, not at startup)
- Extract heartbeat summary functions to infra/heartbeat-summary.ts
  (breaks baileys/jiti/babel chain through auto-reply/reply.ts)
- Remove unnecessary loadPlugins flag from status/health route specs
  (eliminates 2.1MB plugin-loader chunk from startup)

Add dependency-cruiser as devDependency for import chain analysis.

Reduces status command time on VPS from ~9s to ~3.7s (60% faster).
