+++ src/gateway/protocol/version.ts
@@ -1,0 +1,1 @@
+export const PROTOCOL_VERSION = 3 as const;
